Allow two-argument explicit constructor for RequireAlias (#2149)
The latest sync to 703 (https://github.com/luau-lang/luau/pull/2146)
added an explicit constructor to `Luau::RequireAlias`. However, this is
only a single argument constructor for the alias, and does not handle
setting the tags.

This broke luau-lsp compilation of code that did
`Luau::RequireAlias{alias, {"tagname"}}`, and now our only option is to
first construct the RequireAlias into a separate local, and then set the
tags afterwards, then pass the require alias to where we were using it
originally.

For simplicity, this PR adds a two-argument explicit constructor to also
set the tags